A few fixes for gcp.serverless.Function
`@types/express` apparently needs to be a real dependency since users of the library need to be able to load it.

Use real semver version numbers in samples.

My suggestion to use the contents of the deployment time package.json file for the cloud function was misguided, this file is designed for the deployment time program, and will trip up GCF if used at runtime.  Just pass dependencies.

Update licenses to valid SPDX format.