Patch Cloud SQL Database Instance to refresh settings.version
Fixes #794, see also upstream issue:
- https://github.com/hashicorp/terraform-provider-google/issues/15762

Cloud SQL Database Instance resources fail to update in Pulumi when the resource
property `settings.settingsVersion` is modified outside of the Pulumi program.
The API specification requires updates set that property to the current value of
the field. When the value drifts outside of Pulumi's control, the mismatch
results in error 412 on PATCH.

See: https://cloud.google.com/sql/docs/mysql/admin-api/rest/v1/instances#settings

This modifies the resource to always read the current value before performing
the PATCH request that updates the resource.

Testing this posed a few challenges for both writing an upstream acceptance test
and for our own test framework.

For the upstream, the TF SDKv2 test harness has a mandatory `refresh` operation
that runs before applying a configuration. For the upstream issue, a script was
written against the `terraform` CLI to simulate drift, published at
[AaronFriel/google-cloud-sql-instance-412-bug](https://github.com/AaronFriel/google-cloud-sql-instance-412-bug).
That script creates a resource, imports it into a second state and modifies it,
then attempts to update the original state again.

For our own acceptance tests, the challenge was that we lack prior art in this
repository for calling Google Cloud APIs directly. Lacking an API to call to
obtain Google Cloud credentials in the test harness (in the same manner as the
provider does), this PR instead uses the automation API with inline programs to
perform the same set of steps as the test script above.