Add project_name for goreleaser (#1216)
When CI is upgraded to goreleaser v1.21, our publish steps fail as the
`project_name` could not be formulated by GoReleaser. This PR explicitly
sets the `project_name`.

Note, for faster feedback loop, this PR should be merged first and a
follow-up PR in pu/ci-mgmt will be created.
I've also triggered the CI workflow for this commit to test that it
works and it is passing:
https://github.com/pulumi/pulumi-gcp/actions/runs/6313530404/job/17142821989

Fixes: #1217