Rename serviceAccount to serviceaccount (#1305)
Renames the serviceAccount module to serviceaccount according to our convention of lowercase modules.
It's been a long-standing override that we can now get rid of. The change adds aliases to the old type names.

In terms of SDKs, this change is non-breaking for Node, Python, and
.NET, but is breaking for Go and Java. We do need to call it out in our
migration guide.

Once v7 ships, this will fix
https://github.com/pulumi/pulumi-gcp/issues/722