Fix test cleanup for TestRegression794 (#1312)
This PR fixes the clean up steps for `TestRegression794` to prevent
leaking an expensive Cloud SQL DB instance. The 2 changes required are:

1. `DeletionProtection` needs to explicitly be set to `false`
2. The stack backend used for deletion by default attempts to use Pulumi
Cloud. This causes an empty stack to be used for running `pulumi
destroy`, which will not destroy the DB instance. We need to also
specify the `PULUMI_BACKEND_URL` to point to the local backend we've
been using in prior steps

Fixes: #1311