Add auto-naming for gcp:organizations:Project's project_id (#1498)
Fixes https://github.com/pulumi/pulumi-gcp/issues/1492

The `name` field of `gcp:organizations:Project` is cosmetic and does not
need to be unique. It doesn't benefit from a random suffix.

The `project_id` field is structural and globally unique. To allow
replaces (and ease of use), it should be auto-named.