Skip flaky test (#1622)
should fix https://github.com/pulumi/pulumi-gcp/issues/1610

The hypothesis is that the environment variables interact with the ones
from the other parallel tests. That'd explain the non-determinism.

Pretty sure the other tests' environments interact with this one.

I'm going to skip it for now and open an issue about a separate group of
tests which are only run sequentially.