Correct nested labels (#1711)
This PR hardens our labels implementation, ensuring that it correctly
handles arbitrarily nested items.

It incorporates a pre-release version of the bridge to use
https://github.com/pulumi/pulumi-terraform-bridge/pull/1710. Without
#1710, we see a name conflict from the bridge.

Note: `expandPath` should probably be in `pulumi/pulumi`, and I will
open a PR to move it shortly. `resource.PropertyPath.Reset` works
correctly on globed values, but `resource.PropertyPath.Get` and
resource.PropertyPath.Set` do not.


I have been unable to reproduce
https://github.com/pulumi/pulumi-gcp/issues/1387. This PR rewrites
nested label application, which is where the #1387 is broken. I assume
that this will fix it, but in absence of a repro the only proof will be
the absence of new problems.

Fixes https://github.com/pulumi/pulumi-gcp/issues/1387