Rework GCP labels tests to match TF (#2305)
The labels tests were asserting on behaviour different to what TF does.
The previous assumption was that "" in labels means keep old value but
that turns out is not quite correct.

Generally, "" is treated as "keep the previous value". However, if the
value has always been "", then it is treated as "no value". This trumps
any default labels.

This PR reworks the labels tests check to assert on the correct
behaviour, matching TF.

Stacked on https://github.com/pulumi/pulumi-gcp/pull/2277
Part of https://github.com/pulumi/pulumi-gcp/issues/2280