Add regression test for type check warnings (#2389)
This adds a regression test for the warnings we issue to users when a
property has the wrong type. This could result in panics down the line,
so we need to catch them early.

Testing this in the bridge is currently quite hard as we can't use YAML
for this because it is too strict in type-checking. We only have unit
tests which indirectly test this but these break under PRC.

This was introduced in
https://github.com/pulumi/pulumi-terraform-bridge/issues/1987
original GCP issue:
https://github.com/pulumi/pulumi-terraform-bridge/issues/1979

Bridge issue for integration tests for the feature:
https://github.com/pulumi/pulumi-terraform-bridge/issues/2418