Refactor gcp tests to use pulumitest directly (#2508)
The GCP upgrade tests were written in the early days of
[pulumitest](https://github.com/pulumi/providertest/tree/main/pulumitest),
so do not make use of some of the library features. This change
refactors the tests to use it directly. This gets rid of some
boilerplate code and allows direct access to the pulumitest options in
tests.

It also attaches the provider directly which means we no longer have to
keep rebuilding the provider ourselves to test it.

Additionally I've refactored TestCloudRunV2StateMigration which is an
upgrade test with a source edit to use the library code instead of the
code copied from the AWS provider.